requirements.txt
streamlit
pandas
requests
matplotlib
azure-kusto-data
azure-kusto-ingest
azure-identity